requirements.txt
black
slack_bolt
slack_sdk
python-dotenv